+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Menu tab:
         - Don't display logout confirmation when e-mail is set. ([#1684])
+    - Media panel:
+        - Low connection icons. ([#1686])
 
 ### Fixed
 
@@ -28,6 +30,7 @@ All user visible changes to this project will be documented in this file. This p
 [#1680]: /../../pull/1680
 [#1683]: /../../pull/1683
 [#1684]: /../../pull/1684
+[#1686]: /../../pull/1686